Add unit test for exponential backoff duration function
